+++ src/main/java/ti4/discord/interactions/buttons/handlers/draft/FrankenButtonHandler.java
@@ -346,7 +346,9 @@ public class FrankenButtonHandler {
                                     List.of(startStrategyPhaseButton));
                             FrankenDraftBagService.applyDraftBags(event, game, false);
                         } else {
-                            expandFrankenDrazDraft(game, draft);
+                            if (draft instanceof FrankenDrazDraft frankenDrazDraft) {
+                                frankenDrazDraft.expandFactionPackages(game);
+                            }
                             String draftType = "FrankenDraft";
                             if (draft instanceof TwilightsFallFrankenDraft) {
                                 draftType = "Twilight's Fall Draft";
@@ -380,7 +382,9 @@ public class FrankenButtonHandler {
                                         List.of(startStrategyPhaseButton));
                                 FrankenDraftBagService.applyDraftBags(event, game, false);
                             } else {
-                                expandFrankenDrazDraft(game, draft);
+                                if (draft instanceof FrankenDrazDraft frankenDrazDraft) {
+                                    frankenDrazDraft.expandFactionPackages(game);
+                                }
                                 Button randomizeButton =
                                         Buttons.green("startFrankenSliceBuild", "Randomize Your Slices (Sorta)");
                                 Button mantisButton = Buttons.green("startFrankenMantisBuild", "Mantis Build Slices");
@@ -439,36 +443,5 @@ public class FrankenButtonHandler {
 
         FrankenDraftBagService.showPlayerBag(game, player);
         event.getMessage().delete().queue(Consumers.nop(), BotLogger::catchRestError);
-    }
-
-    @ButtonHandler(value = "frankenFactionComponents", save = false)
-    private static void showFactionComponents(ButtonInteractionEvent event, Player player, String buttonID) {
-        if (player.getGame().getActiveBagDraft() instanceof FrankenDrazDraft frankenDrazDraft) {
-            String faction = buttonID.split(";")[1];
-            frankenDrazDraft.sendFactionComponentCards(event, player, faction);
-        }
-    }
-
-    @ButtonHandler(value = "frankenDrazCategory", save = false)
-    private static void showFrankenDrazCategory(ButtonInteractionEvent event, Player player, String buttonID) {
-        if (player.getGame().getActiveBagDraft() instanceof FrankenDrazDraft frankenDrazDraft) {
-            DraftCategory category = DraftCategory.valueOf(buttonID.split(";")[1]);
-            frankenDrazDraft.sendPostDraftCategory(player, category);
-        }
-    }
-
-    @ButtonHandler(value = "frankenDrazCloseCategory", save = false)
-    private static void closeFrankenDrazCategory(ButtonInteractionEvent event, Player player, String buttonID) {
-        if (player.getGame().getActiveBagDraft() instanceof FrankenDrazDraft frankenDrazDraft) {
-            DraftCategory category = DraftCategory.valueOf(buttonID.split(";")[1]);
-            frankenDrazDraft.closePostDraftCategory(event, player, category);
-            MessageHelper.sendEphemeralMessageToEventChannel(event, "Closed " + category.toString() + ".");
-        }
-    }
-
-    private static void expandFrankenDrazDraft(Game game, BagDraft draft) {
-        if (draft instanceof FrankenDrazDraft frankenDrazDraft) {
-            frankenDrazDraft.expandFactionPackages(game);
-        }
     }
 }
